move bible-books as a seperated package
