+++ packages/bible-books/data/translations-non-western/README.md
