+++ packages/bible-books/data/translations.test.ts
