+++ packages/bible-books/data/translations/README.md
