+++ packages/bible-books/data/translations/en.json
