+++ packages/bible-books/data/translations/it.json
