+++ packages/bible-books/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "bible-books",
+  "version": "0.0.0",
+  "main": "dist/index.js",
+  "devDependencies": {},
+  "scripts": {
+    "build": "tsc"
+  }
+}
